SCSS Overhaul Conversion!!!  (one of many)
Overhaul the styling of the system to make use of Bourbon.io+bitters+neat.
Involves grouping form elements into a common div on the form section (as opposed to being on separate lines).
Remove the sexy oop-oop-oops in favour of a more maintainable and logical approach, (the name of the element that's being surrounded!).